+++ src/Services/Catalog/CatalogServiceBuilder.php
@@ -118,4 +118,68 @@ class CatalogServiceBuilder extends AbstractServiceBuilder
 
         return $this->serviceCache[__METHOD__];
     }
+
+    public function price(): Catalog\Price\Service\Price
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Catalog\Price\Service\Price(
+                new Catalog\Price\Service\Batch(
+                    new Catalog\Price\Batch($this->core, $this->log),
+                    $this->log
+                ),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function priceType(): Catalog\PriceType\Service\PriceType
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Catalog\PriceType\Service\PriceType(
+                new Catalog\PriceType\Service\Batch(
+                    new Catalog\PriceType\Batch($this->core, $this->log),
+                    $this->log
+                ),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function priceTypeLang(): Catalog\PriceTypeLang\Service\PriceTypeLang
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Catalog\PriceTypeLang\Service\PriceTypeLang(
+                new Catalog\PriceTypeLang\Service\Batch(
+                    new Catalog\PriceTypeLang\Batch($this->core, $this->log),
+                    $this->log
+                ),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function priceTypeGroup(): Catalog\PriceTypeGroup\Service\PriceTypeGroup
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Catalog\PriceTypeGroup\Service\PriceTypeGroup(
+                new Catalog\PriceTypeGroup\Service\Batch(
+                    new Catalog\PriceTypeGroup\Batch($this->core, $this->log),
+                    $this->log
+                ),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
 }