+++ src/Services/Catalog/Price/Result/PricesResult.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Dmitriy Ignatenko <algonexys@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Catalog\Price\Result;
+
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Result\AbstractResult;
+
+class PricesResult extends AbstractResult
+{
+    /**
+     * @return PriceItemResult[]
+     * @throws BaseException
+     */
+    public function getPrices(): array
+    {
+        $result = $this->getCoreResponse()->getResponseData()->getResult();
+
+        return array_map(
+            static fn (array $item): PriceItemResult => new PriceItemResult($item),
+            $result['prices'] ?? []
+        );
+    }
+}